% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUCDiffP}
\alias{AUCDiffP}
\title{Generate Permutation P-value for the Difference in AUCs}
\usage{
AUCDiffP(time, status, arm, tau, B = 2000)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{tau}{Truncation time.}

\item{B}{Permutations.}
}
\value{
Numeric vector containing the truncation time `Time`, the
  estimated AUCs in arms 1 `Arm1` and 0 `Arm0`, the difference `Delta`,
  and the permutation p-value `P`.
}
\description{
Permutation p-value for the difference in areas under the
cumulative incidence curves.
}
