% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUCIC.Stats}
\alias{AUCIC.Stats}
\title{Calculate Difference and Ratio in AUCs.}
\usage{
AUCIC.Stats(data, tau, return_areas = FALSE)
}
\arguments{
\item{data}{Data.frame containing `time`, `status`, `arm`.}

\item{tau}{Truncation time.}

\item{return_areas}{Return the AUCs?}
}
\value{
If `return_areas = TRUE`, list containing:
\itemize{
  \item `areas` under the cumulative incidence curve for each arm.
  \item `stats`, including the difference and ratio of areas.
}
 If `return_areas = FALSE`, only `stats` is returned.
}
\description{
Calculate Difference and Ratio in AUCs.
}
