% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{BootCI}
\alias{BootCI}
\title{Construct Confidence Interval}
\usage{
BootCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Numeric vector containing the 'alpha' level, the 'lower' and 'upper'
  confidence bounds.
}
\description{
Construct Confidence Interval
}
