% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{BootData}
\alias{BootData}
\title{Generate Bootstrap Data.frame}
\usage{
BootData(data)
}
\arguments{
\item{data}{Data.frame.}
}
\value{
Bootstrapped data.frame.
}
\description{
Generate Bootstrap Data.frame
}
