% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIC.R
\name{CalcCIC}
\alias{CalcCIC}
\title{Calculate Cumulative Incidence Curve.}
\usage{
CalcCIC(time, status)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}
}
\value{
Tabulate cumulative incidence curve.
}
\description{
Calculate Cumulative Incidence Curve.
}
