% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CICs-package.R
\docType{package}
\name{CICs-package}
\alias{CICs}
\alias{CICs-package}
\title{CICs: Inference on Cumulative Incidence Curves}
\description{
Provides functions for inference on the difference and ratio in event rates, quantiles, and areas under the cumulative incidence curves, comparing two treatment arms.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
