% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{CompareCICs}
\alias{CompareCICs}
\title{Inference on CICs.}
\usage{
CompareCICs(
  time,
  status,
  arm,
  sum_stat = "AUC",
  param = NULL,
  reps = 2000,
  alpha = 0.05
)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{sum_stat}{Summary statistic, from among 'AUC', 'Quantile', 'Rate'.}

\item{param}{Either the truncation time, if `sum_stat` is 'AUC' or 'Rate', or 
the quantile probability, if `sum_stat` is 'Quantile'.}

\item{reps}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Object of class \code{compCICs} containing:
\itemize{
  \item `@CIs`: Observed difference and ratio in areas with confidence intervals. 
  \item `@Curves`: Cumulative incidence curves. 
  \item `@Pvals`: Bootstrap and permutation p-values.
  \item `@Reps`: Bootstrap and permutation realizations of the test statistics. 
  \item `@Stats`: Observed test statistics. 
}
}
\description{
Inference for the difference and ratio in summary statistics comparing
two CICs.
}
