% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event-rates.R
\name{CompareEventRates}
\alias{CompareEventRates}
\title{Bootstrap Inference for the Event Rate Difference.}
\usage{
CompareEventRates(time, status, arm, tau, reps = 2000, alpha = 0.05)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{tau}{Event time.}

\item{reps}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Data.frame containing these columns:
\describe{
  \item{Time}{Truncation time.}
  \item{Arm0}{Event rate for arm 0.}
  \item{Arm1}{Event rate for arm 1.}
  \item{Contrast}{'A1-A0' for the difference and 'A1/A0' for the ratio.}
  \item{Estimate}{The estimated difference and ratio of AUCs.}
  \item{L}{Confidence lower bound.}
  \item{U}{Confidence upper bound.}
  \item{P}{P-value.}
}
}
\description{
Bootstrap inference for the difference and ratio of event rates
at time tau.
}
