% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{compCICs-class}
\alias{compCICs-class}
\title{Compare CICs Object}
\description{
Defines the object class returned by \code{\link{CompareCICs}}.
}
\section{Slots}{

\describe{
\item{\code{cis}}{Confidence intervals.}

\item{\code{cics}}{Cumulative incidence curves.}

\item{\code{pvals}}{Bootstrap and permutation p-values.}

\item{\code{reps}}{Bootstrap and permutation statistics.}

\item{\code{marg_stats}}{Observed test statistics.}

\item{\code{stratum_stats}}{Per-stratum weights and statistics.}
}}

