% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIs.R
\name{ConstructCI}
\alias{ConstructCI}
\title{Construct Confidence Interval}
\usage{
ConstructCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Numeric vector containing the `Alpha` level, and the 
  lower `L` and upper `U` confidence bounds.
}
\description{
Construct Confidence Interval
}
