% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event-rates.R
\name{EventRate.Stats}
\alias{EventRate.Stats}
\title{Calculate Difference and Ratio in Event Rates.}
\usage{
EventRate.Stats(data, tau, return_rates = FALSE)
}
\arguments{
\item{data}{Data.frame containing `time`, `status`, `arm`.}

\item{tau}{Truncation time.}

\item{return_rates}{Return the event rates?}
}
\value{
If `return_rates = TRUE`, list containing:
\itemize{
  \item `rates` for each arm. 
  \item `stats`, including the difference and ratio of areas.
}
 If `return_rates = FALSE`, only `stats` is returned.
}
\description{
Calculate Difference and Ratio in Event Rates.
}
