% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event-rates.R
\name{EventRateDiffP}
\alias{EventRateDiffP}
\title{Generate Permutation P-value for the Event Rate Difference.}
\usage{
EventRateDiffP(time, status, arm, tau, B = 2000)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{tau}{Event time.}

\item{B}{Permutations.}
}
\value{
Numeric vector containing the event time `Time`, the estimated rates
  in arms 1 `Arm1` and 0 `Arm0`, the difference `Delta`, and the permutation
  p-value `P`.
}
\description{
Permutation p-value the difference in event rates at time time tau.
}
