% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{FindAOC}
\alias{FindAOC}
\title{Find Area Over the Cumulative Incidence Curve.}
\usage{
FindAOC(times, probs, tau)
}
\arguments{
\item{times}{CIC times.}

\item{probs}{CIC probabilities.}

\item{tau}{Truncation time.}
}
\description{
Find Area Over the Cumulative Incidence Curve.
}
