% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{FindQuantile}
\alias{FindQuantile}
\title{Find Quantile of Cumulative Incidence Curve.}
\usage{
FindQuantile(times, probs, q)
}
\arguments{
\item{times}{CIC times.}

\item{probs}{CIC probabilities.}

\item{q}{Quantile.}
}
\description{
Find Quantile of Cumulative Incidence Curve.
}
