% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{FindRate}
\alias{FindRate}
\title{Find Event Rate at Time tau.}
\usage{
FindRate(times, probs, tau)
}
\arguments{
\item{times}{CIC times.}

\item{probs}{CIC probabilities.}

\item{tau}{Event time.}
}
\description{
Find Event Rate at Time tau.
}
