% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{FindStat}
\alias{FindStat}
\title{Find Summary Statistic}
\usage{
FindStat(times, probs, sum_stat, param)
}
\arguments{
\item{times}{CIC times.}

\item{probs}{CIC probabilities.}

\item{sum_stat}{Summary statistic, from among 'AUC', 'Quantile', 'Rate'.}

\item{param}{Either the truncation time, if `sum_stat` is 'AUC' or 'Rate', or 
the quantile probability, if `sum_stat` is 'Quantile'.}
}
\value{
Numeric summary statistic.
}
\description{
Calculate a summary statistic from a cumulative incidence curve.
}
