% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,compCICs-method}
\alias{show,compCICs-method}
\title{Show Method for Compare CICs Object}
\usage{
\S4method{show}{compCICs}(object)
}
\arguments{
\item{object}{An object of class \code{compCICs}.}
}
\description{
Show Method for Compare CICs Object
}
