% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Generate Competing Risks Data}
\usage{
GenData(n, event_rate, death_rate, censor_rate = 0, tau = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{event_rate}{Event rate.}

\item{death_rate}{Death rate.}

\item{censor_rate}{Censoring rate.}

\item{tau}{Optional truncation time.}
}
\description{
Assumes exponential cause-specific hazards and independent exponential
censoring.
}
