% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{GenTwoSampleData}
\alias{GenTwoSampleData}
\title{Generate Two-Sample Data}
\usage{
GenTwoSampleData(
  n1,
  n0,
  censor_rate1 = 0.25,
  censor_rate0 = 0.25,
  event_rate1 = 0.5,
  event_rate0 = 0.5,
  death_rate1 = 0.25,
  death_rate0 = 0.25,
  tau = NULL
)
}
\arguments{
\item{n1}{Sample size in arm 1.}

\item{n0}{Sample size in arm 0.}

\item{censor_rate1}{Censoring rate in arm 1.}

\item{censor_rate0}{Censoring rate in arm 0.}

\item{event_rate1}{Event rate in arm 1.}

\item{event_rate0}{Event rate in arm 0.}

\item{death_rate1}{Death rate in arm 1.}

\item{death_rate0}{Death rate in arm 0.}

\item{tau}{Optional truncation time needed if censoring rate is set to 0.}
}
\value{
Data.frame with observation 'time', event 'status, and 
  treatment 'arm'.
}
\description{
Assumes exponential cause-specific hazards and independent exponential
censoring.
}
