% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{OneSampleNARDF}
\alias{OneSampleNARDF}
\title{One-Sample Number at Risk Plotting Frame}
\usage{
OneSampleNARDF(data, x_breaks, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{Time points at which to determine the NARs.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{arm_name}{Name of arm column.}
}
\value{
Data.frame containing `time`, `nar_ctrl`, `nar_trt`.
}
\description{
One-sample numbers at risk for competing risks data.
}
