% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{PermData}
\alias{PermData}
\title{Generate Permuted Data.frame}
\usage{
PermData(data)
}
\arguments{
\item{data}{Data.frame containing `arm`.}
}
\value{
Permuted data.frame.
}
\description{
Generate Permuted Data.frame
}
