% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\name{PermSim}
\alias{PermSim}
\title{Permutation Inference.}
\usage{
PermSim(data, obs_stats, sum_stat, param, alpha, reps)
}
\arguments{
\item{data}{Data.frame containing: time, status, arm, strata.}

\item{obs_stats}{Observed contrasts.}

\item{sum_stat}{Summary statistic.}

\item{param}{Parameter.}

\item{alpha}{Type I error.}

\item{reps}{Simulations replicates.}
}
\value{
List
}
\description{
Permutation Inference.
}
