% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotCICs}
\alias{PlotCICs}
\title{Plot Cumulative Incidence Curves}
\usage{
PlotCICs(
  data,
  arm_name = "arm",
  color_labs = c("Ctrl", "Trt"),
  ctrl_color = "#C65842",
  legend_pos = "top",
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  trt_color = "#6385B8",
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Cumulative Incidence",
  y_lim = c(0, 1)
)
}
\arguments{
\item{data}{Data.frame.}

\item{arm_name}{Name of arm column.}

\item{color_labs}{Color labels.}

\item{ctrl_color}{Color for control arm.}

\item{legend_pos}{Legend position.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{trt_color}{Color for treatment arm.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit, may differ from tau.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot the cumulative incidence curves comparing two treatment arms.
}
