% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotOneSampleCIC}
\alias{PlotOneSampleCIC}
\title{Plot One-Sample Cumulative Incidence Curve}
\usage{
PlotOneSampleCIC(
  data,
  color = "#C65842",
  status_name = "status",
  tau = NULL,
  time_name = "time",
  title = NULL,
  x_breaks = NULL,
  x_labs = NULL,
  x_name = "Time",
  x_max = NULL,
  y_name = "Cumulative Incidence",
  y_lim = c(0, 1)
)
}
\arguments{
\item{data}{Data.frame.}

\item{color}{Line color.}

\item{status_name}{Name of status column.}

\item{tau}{Truncation time.}

\item{time_name}{Name of time column.}

\item{title}{Plot title.}

\item{x_breaks}{X-axis breaks.}

\item{x_labs}{X-axis labels.}

\item{x_name}{X-axis name.}

\item{x_max}{X-axis upper limit, may differ from tau.}

\item{y_name}{Y-axis name.}

\item{y_lim}{Y-axis limits.}
}
\value{
ggplot.
}
\description{
Plot the cumulative incidence curve for a single treatment arm.
}
