% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotOneSampleNARs}
\alias{PlotOneSampleNARs}
\title{Plot One-Sample Numbers at Risk}
\usage{
PlotOneSampleNARs(
  data,
  x_breaks,
  status_name = "status",
  time_name = "time",
  x_labs = NULL,
  x_max = NULL,
  x_name = NULL,
  y_lab = "NAR"
)
}
\arguments{
\item{data}{Data.frame.}

\item{x_breaks}{X-axis breaks.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}

\item{x_labs}{X-axis labels.}

\item{x_max}{X-axis upper limit.}

\item{x_name}{X-axis name.}

\item{y_lab}{Y-axis labels.}
}
\value{
ggplot.
}
\description{
Plot One-Sample Numbers at Risk
}
