% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.compCICs}
\alias{print.compCICs}
\title{Print Method for Compre CICs Object.}
\usage{
\method{print}{compCICs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{compCICs}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{compCICs}.
}
