% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{QuantDiffCI}
\alias{QuantDiffCI}
\title{Generate Bootstrap Confidence Interval for the Difference in Quantiles.}
\usage{
QuantDiffCI(time, status, arm, q = 0.5, B = 2000, alpha = 0.05)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}

\item{arm}{Arm, assumed to have two levels, coded 0/1.}

\item{q}{Quantile probability.}

\item{B}{Bootstrap replicates.}

\item{alpha}{Alpha level.}
}
\value{
Numeric vector containing the quantile probability `Quantile`, the
  estimated quantiles in arms 1 `Arm1` and 0 `Arm0`, the difference `Delta`,
  and the lower `L` and upper `U` confidence bounds.
}
\description{
Bootstrap confidence interval for the difference in quantiles.
}
