% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantiles.R
\name{Quantile.Stats}
\alias{Quantile.Stats}
\title{Calculate Difference and Ratio in Event Rates.}
\usage{
Quantile.Stats(data, q, return_quants = FALSE)
}
\arguments{
\item{data}{Data.frame containing `time`, `status`, `arm`.}

\item{q}{Quantile probability.}

\item{return_quants}{Return the quantiles?}
}
\value{
If `return_quants = TRUE`, list containing:
\itemize{
  \item `rates` for each arm. 
  \item `stats`, including the difference and ratio of areas.
}
 If `return_quants = FALSE`, only `stats` is returned.
}
\description{
Calculate Difference and Ratio in Event Rates.
}
