% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{SECurve}
\alias{SECurve}
\title{Standard Error of Cumulative Incidence Curve.}
\usage{
SECurve(data, status_name = "status", time_name = "time")
}
\arguments{
\item{data}{Data.frame.}

\item{status_name}{Name of status column.}

\item{time_name}{Name of time column.}
}
\value{
stepfun.
}
\description{
Return a function that calculates the standard error of the event probability
for a single treatment arm. Assumes status == 1 is the event of interest.
}
