% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{StratPerm}
\alias{StratPerm}
\title{Stratified Permutation}
\usage{
StratPerm(data)
}
\arguments{
\item{data}{Data.frame containing `strata` and `arm`.}
}
\value{
Data.frame bootstrapped within strata.
}
\description{
Stratified Permutation
}
