% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{SumStats}
\alias{SumStats}
\title{Calculate Difference and Ratio in Summary Stats}
\usage{
SumStats(data0, data1, sum_stat = "AUC", param, return_per_arm = FALSE)
}
\arguments{
\item{data0}{Data.frame containing `time` and `status` for arm 0.}

\item{data1}{Data.frame containing `time` and `status` for arm 1.}

\item{sum_stat}{Summary statistic, from among 'AUC', 'Quantile', 'Rate'.}

\item{param}{Either the truncation time, if `sum_stat` is 'AUC' or 'Rate', or 
the quantile probability, if `sum_stat` is 'Quantile'.}

\item{return_per_arm}{Return per-Arm stats and CICs?}
}
\value{
If `return_per_arm == TRUE`, list containing:
\itemize{
  \item `cic`, tabulated cumulative incidence curves.
  \item `per_arm`, the per-arm summary statistics. 
  \item `stats`, including the difference and ratio of summary statistics. 
}
 If `return_per_arm == FALSE`, only `stats` is returned.
}
\description{
Calculate Difference and Ratio in Summary Stats
}
