% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabulate-CIC.R
\name{TabulateCIC}
\alias{TabulateCIC}
\title{Tabulate CIC}
\usage{
TabulateCIC(time, status)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Event status. The event coded as 1 is assumed to be the event
of interest.}
}
\value{
Data.frame containing `Time`, the estimated `CIC`, the variance `Var_CIC` of the
  CIC, and event `Status`.
}
\description{
Converts the list returned by \code{\link[cmprsk]{cuminc}} into a table.
}
