% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{compCICs-class}
\alias{compCICs-class}
\title{Compare CICs Object}
\description{
Defines the object class returned by \code{\link{CompareCICs}}.
}
\section{Slots}{

\describe{
\item{\code{CIs}}{Confidence intervals.}

\item{\code{Curves}}{Cumulative incidence curves.}

\item{\code{Pvals}}{Bootstrap and permutation p-values.}

\item{\code{Reps}}{Bootstrap and permutation statistics.}

\item{\code{Stats}}{Observed test statistics.}
}}

