% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{FindAUC}
\alias{FindAUC}
\title{Find Area Under the Cumulative Incidence Curve.}
\usage{
FindAUC(times, probs, tau)
}
\arguments{
\item{times}{CIC times.}

\item{probs}{CIC probabilities.}

\item{tau}{Truncation time.}
}
\value{
Numeric.
}
\description{
Find Area Under the Cumulative Incidence Curve.
}
