% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenData.R
\name{GenData}
\alias{GenData}
\title{Generate Competing Risks Data}
\usage{
GenData(n, censor_rate, event_rate, death_rate, tau = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{censor_rate}{Censoring rate.}

\item{event_rate}{Event rate.}

\item{death_rate}{Death rate.}

\item{tau}{Optional truncation time needed if censoring rate is set to 0.}
}
\value{
Data.frame containing the observation 'time' and event 'status'.
}
\description{
Assumes exponential cause-specific hazards and independent exponential
censoring.
}
