% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{StratBoot}
\alias{StratBoot}
\title{Stratified Bootstrap}
\usage{
StratBoot(data)
}
\arguments{
\item{data}{Data.frame containing `strata`}
}
\value{
Data.frame bootstrapped within strata.
}
\description{
Stratified Bootstrap
}
