% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{SumStats}
\alias{SumStats}
\title{Calculate Difference and Ratio in Summary Stats}
\usage{
SumStats(data, sum_stat = "AUC", param, return_strata = FALSE)
}
\arguments{
\item{data}{Data.frame containing 'time', 'status', 'arm', 'strata'.}

\item{sum_stat}{Summary statistic, from among 'AOC', 'AUC', 'Quantile', 'Rate'.}

\item{param}{Truncation time, if `sum_stat` is 'AOC', 'AUC', 'Rate';
the quantile probability, if `sum_stat` is 'Quantile'.}

\item{return_strata}{Return per_stratum stats and CICs?}
}
\value{
If `return_per_arm`, list containing:
\itemize{
  \item {`contrasts`: difference and ratio of summary statistics.} 
  \item {`curves`: list of tabulated cumulative incidence curves.}
  \item {`marg`: marginal summary statistics.}
  \item {`weights`: per-stratum summary statistics.}
}
}
\description{
Calculate Difference and Ratio in Summary Stats
}
