% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabulate-CIC.R
\name{TabulateCIC}
\alias{TabulateCIC}
\title{Tabulate CIC}
\usage{
TabulateCIC(fit)
}
\arguments{
\item{fit}{Fitted object of class "cuminc".}
}
\value{
Data.frame containing `Time`, the estimated `CIC`, the variance `Var_CIC` of the
  CIC, and event `Status`.
}
\description{
Converts from the list returned by \code{\link[cmprsk]{cuminc}}.
}
